refactor(reposicao): split AdminReposicaoNegociacaoParalela (702→280)
Extrai a camada de dados/lógica para o hook useNegociacaoParalela
(estado, 2 queries, memos derivados e handlers) e quebra o JSX em
componentes presentacionais controlados (SugestoesToolbar, RankingToolbar,
DistribuicaoCards, RankingPaginacao). toggleSet movido para helpers.
Comportamento preservado 1:1; +16 testes (componentes puros + toggleSet).
